A2

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: abouchfa <abouchfa@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/01 13:08:37 by abouchfa          #+#    #+#             */
-/*   Updated: 2021/11/12 16:23:40 by abouchfa         ###   ########.fr       */
+/*   Updated: 2021/11/14 15:10:57 by abouchfa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,7 +32,7 @@ int	ft_atoi(const char *str)
 		n = n * 10 + (*str - 48);
 		str++;
 	}
-	return ((int) n * minus);
+	return (n * minus);
 }
 
 // int main(int ac, char **av)
